addec receiving energy from gnerator

+++ src/main/java/com/craftle_mod/common/tile/storage/energy_matrix/EnergyMatrixTileEntityFactory.java
@@ -10,8 +10,8 @@ import java.util.function.Supplier;
 public class EnergyMatrixTileEntityFactory implements TileEntityFactory {
 
     public Supplier<EnergyMatrixTileEntity> buildSupplier(CraftleBaseTier tier) {
-        long capacity =
-                (long) (TileEntityConstants.ENERGY_MATRIX_BASE_CAPACITY * tier.getMultiplier());
+        int capacity =
+                (int) (TileEntityConstants.ENERGY_MATRIX_BASE_CAPACITY * tier.getMultiplier());
         switch (tier) {
             case TIER_1:
                 return (() -> new EnergyMatrixTileEntity(
